flask-app-fido/requirements.txt
Flask==2.2.3
Authlib==1.2.1
requests==2.28.2
PyJWT==2.6.0
werkzeug==2.2.3 